Create a safe way to invoke methods on Mongo
