--- backend_app/database/schemas/PingSchema.js
@@ -1,11 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const Ping = new Schema(
-  {
-    body: { type: String, required: true },
-  },
-  { timestamps: true }
-);
-
-module.exports = mongoose.model("pings", Ping);
